Adapt tests for refactored build description
