+++ lib/graph/Module.js
@@ -234,7 +234,12 @@ class Module {
 	async _getSuppliedConfigurations() {
 		if (this._suppliedConfigs.length) {
 			log.verbose(`Configuration for module ${this.getId()} has been supplied directly`);
-			return await Promise.all(this._suppliedConfigs.map(async (config) => {
+			return await Promise.all(this._suppliedConfigs.map(async (suppliedConfig) => {
+				let config = suppliedConfig;
+				if (suppliedConfig.buildDescription) {
+					config = suppliedConfig.project;
+					config._buildDescription = suppliedConfig.buildDescription;
+				}
 				return this._normalizeAndApplyShims(config);
 			}));
 		}
@@ -354,15 +359,15 @@ class Module {
 	}
 
 	async _getBuildDescriptionConfigurations() {
-		const cacheConfig = await this._readBuildDescription();
+		const buildDescriptionMetadata = await this._readBuildDescription();
 
-		if (!cacheConfig) {
+		if (!buildDescriptionMetadata) {
 			log.verbose(`Could not find any build description in module ${this.getId()}`);
 			return [];
 		}
 
-		const config = cacheConfig.project;
-		config._buildDescription = cacheConfig.buildDescription;
+		const config = buildDescriptionMetadata.project;
+		config._buildDescription = buildDescriptionMetadata.buildDescription;
 		return [this._normalizeAndApplyShims(config)];
 	}
