+++ test/fixtures/build-descriptions/application.a/package.json
