+++ test/fixtures/build-descriptions/application.a/resources/id1/manifest.json
