+++ test/fixtures/build-descriptions/application.a/resources/id1/test-dbg.js
