+++ test/fixtures/build-descriptions/library.e/package.json
