+++ test/fixtures/build-descriptions/library.e/resources/library/e/.library
